[p1p2] in gg_tt01g.sa, fix builds in P2 as done previously in P1 for thw two process-dependent files CPPProcess.cc and check_sa.cc

+++ epochX/cudacpp/gg_tt01g.sa/SubProcesses/P1_Sigma_sm_gg_ttx/CPPProcess.cc
@@ -108,7 +108,7 @@ namespace mg5amcCpu
   // In CUDA, this device function computes the ME for a single event
   // In C++, this function computes the ME for a single event "page" or SIMD vector (or for two in "mixed" precision mode, nParity=2)
   __device__ INLINE void /* clang-format off */
-  calculate_wavefunctions( int ihel,
+  calculate_wavefunctions( size_t ihel,
                            const fptype* allmomenta,      // input: momenta[nevt*npar*4]
                            const fptype* allcouplings,    // input: couplings[nevt*ndcoup*2]
                            fptype* allMEs,                // output: allMEs[nevt], |M|^2 running_sum_over_helicities
